fixup: Move util functions to utils, update types and typos
